+++ src/context/FeatureEnable/FeatureEnableContext.tsx
@@ -3,9 +3,11 @@ import { createContext, Dispatch, SetStateAction } from "react";
 export interface DefaultValueFeatureEnableContext {
   features?: string[];
   setFeatures: Dispatch<SetStateAction<string[] | undefined>>;
+  disabledFeatures?: string[];
 }
 
 export const defaultValueFeatureEnableContext: DefaultValueFeatureEnableContext = {
+  disabledFeatures: undefined,
   features: undefined,
   setFeatures: () => {},
 };