+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v1.37.2
-- **[build]** - fix some type export
+## v1.38.0
+- **[feat]** - add `disabledFeatures` props to `FeatureProvider`
+- **[enhancement]** - `useFeatureEnabled` hook